update configuration to use new connectionstring parser
