+++ src/RabbitMQ/NServiceBus.RabbitMQ/NServiceBus.Transports.RabbitMQ.csproj
@@ -72,8 +72,6 @@
     <Compile Include="EasyNetQ\PersistentConnection.cs" />
     <Compile Include="IManageRabbitMqConnections.cs" />
     <Compile Include="RabbitMqConnectionManager.cs" />
-    <Compile Include="Config\RabbitMqTransportConfigurer.cs" />
-    <Compile Include="Config\RabbitMqConnectionStringParser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RabbitMqDequeueStrategy.cs" />
     <Compile Include="RabbitMqMessagePublisher.cs" />
